Adding docker information for CNV analysis
I made the docker image for reproducibility of the CNV figures. So the docker image includes specific reference genome fasta files making the docker build folder too big. So I am adding the dockerhub link in the Readme instead of pushing the docker build folder to github. Currently the docker image exists in my personal dockerhub. Is it recommended to transfer it to nfosi?